chore(mobile): Code tidy Messages / Thread area and add notes
